requirements.txt
flask
pymupdf
python-docx
termcolor